backend/requirements.txt
fastapi==0.95.0
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.21
alembic==1.11.1
psycopg2-binary==2.9.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==1.10.12
celery==5.3.1
redis==4.5.5
python-dotenv==1.0.0
aiofiles==23.1.0
httpx==0.24.1
email-validator==2.0.0
python-multipart==0.0.6
